refactor: move components in Daily page to separate folder
- Refactor directory structure for organization
