+++ frontend/src/Daily/components/DateNav.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useRecoilState } from 'recoil';
 import { IconButton } from '@material-ui/core';
 import { ChevronLeft, ChevronRight } from '@material-ui/icons';
-import { dateState } from './states';
+import { dateState } from '../states';
 
 const DAY_MILLISEC = 1000 * 60 * 60 * 24;
