+++ frontend/src/Daily/components/WorkoutList.tsx
@@ -3,8 +3,8 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 import { Grid } from '@material-ui/core';
 
 import Workout from './Workout';
-import { dateState, getWorkoutsQuery } from './states';
-import { getDateString } from './services';
+import { dateState, getWorkoutsQuery } from '../states';
+import { getDateString } from '../services';
 
 function WorkoutList() {
   const [date] = useRecoilState(dateState);
